+++ CHANGELOG.md
@@ -5,6 +5,7 @@ HEAD
 ----
 
 * Fix reconfigure NAPT after IP_EVENT_STA_GOT_IP for esp32
+* Replace LinkedList with std::vector. Fix issue #8
 
 v2.2.2 (2023-05-19)
 ------